+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 - Added passwordstore support for storing passwords (the linux 'pass' command
   line tool)
+- Added **content-edit-select-all** in content edit mode, bound to **C-x h**.
 
 ### Changed